WIP: varargs mechanism overhaul
