+++ src/main/java/com/github/dakusui/scriptiveunit/model/form/handle/FormHandleFactory.java
@@ -1,9 +1,7 @@
 package com.github.dakusui.scriptiveunit.model.form.handle;
 
 import com.github.dakusui.scriptiveunit.core.Config;
-import com.github.dakusui.scriptiveunit.core.ObjectMethod;
 import com.github.dakusui.scriptiveunit.model.form.Form;
-import com.github.dakusui.scriptiveunit.model.form.FormUtils;
 import com.github.dakusui.scriptiveunit.model.statement.Statement;
 import com.github.dakusui.scriptiveunit.utils.DriverUtils;
