+++ src/main/java/com/github/dakusui/scriptiveunit/model/form/handle/FormRegistry.java
@@ -2,7 +2,6 @@ package com.github.dakusui.scriptiveunit.model.form.handle;
 
 import com.github.dakusui.jcunit.core.tuples.Tuple;
 import com.github.dakusui.scriptiveunit.core.Config;
-import com.github.dakusui.scriptiveunit.core.ObjectMethod;
 import com.github.dakusui.scriptiveunit.drivers.actions.Basic;
 import com.github.dakusui.scriptiveunit.model.desc.testitem.TestItem;
 import com.github.dakusui.scriptiveunit.model.form.Form;
