+++ src/test/java/com/github/dakusui/scriptiveunit/testassets/Driver1.java
@@ -1,16 +1,18 @@
 package com.github.dakusui.scriptiveunit.testassets;
 
+import com.github.dakusui.scriptiveunit.annotations.Doc;
+import com.github.dakusui.scriptiveunit.annotations.Import;
 import com.github.dakusui.scriptiveunit.annotations.Import.Alias;
+import com.github.dakusui.scriptiveunit.annotations.Scriptable;
 import com.github.dakusui.scriptiveunit.drivers.Arith;
 import com.github.dakusui.scriptiveunit.drivers.Predicates;
 import com.github.dakusui.scriptiveunit.drivers.actions.Basic;
-import com.github.dakusui.scriptiveunit.runners.ScriptiveUnit;
-import com.github.dakusui.scriptiveunit.annotations.Doc;
-import com.github.dakusui.scriptiveunit.annotations.Import;
-import com.github.dakusui.scriptiveunit.annotations.Scriptable;
-import com.github.dakusui.scriptiveunit.model.session.Stage;
 import com.github.dakusui.scriptiveunit.model.form.Form;
+import com.github.dakusui.scriptiveunit.model.session.Stage;
+import com.github.dakusui.scriptiveunit.runners.ScriptiveUnit;
 import org.junit.runner.RunWith;
+
+import java.util.List;
 
 @RunWith(ScriptiveUnit.class)
 public class Driver1 {
@@ -28,8 +30,12 @@ public class Driver1 {
       @Alias(value = "eq", as = "=="),
       @Alias(value = "ifthen", as = "if_then")
   })
+  public final Object predicates = new Predicates();
 
-  public final Object arith = new Predicates();
+  @SuppressWarnings("unused")
+  @Import
+  public final Object arith = new Arith();
+
   @SuppressWarnings("unused")
   @Import
   public final Object helloWorld = new HelloWorld();
@@ -48,6 +54,19 @@ public class Driver1 {
       return (Stage input) -> {
         System.out.println(s.apply(input));
         return s.apply(input);
+      };
+    }
+
+    @SuppressWarnings("unused")
+    @Scriptable
+    public Form<String> printVarargs(List<Form<String>> strings) {
+      StringBuffer buf = new StringBuffer();
+      return (Stage input) -> {
+        strings.forEach(s -> {
+          System.out.println(s.apply(input));
+          buf.append(s.apply(input));
+        });
+        return buf.toString();
       };
     }
   }
